migrate the IPAM information from host-local to whereabouts CRD
this commit provides an upgrade path for exisitng zones using host-local
IPAM for sriov public subnet. when the whereabouts daemonset is installed,
the entrypoint script for the container calls `migrate` utility. This
utility looks at IPAM results files in /var/lib/cni/results directory and
for all the sriov public network files, it reserves the IP that is present
in that file.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>